Begin LWIP integration

+++ include/stddef.h
@@ -6,7 +6,7 @@
 #define NULL 0
 #define offsetof(st, m) __builtin_offsetof(st, m)
 
-typedef int ptrdiff_t;
+typedef int64_t ptrdiff_t;
 typedef uint16_t wchar_t;
 typedef uint64_t size_t;
